Copy example docs from the Main VIs to README files
This patch copies (with minor edits) the documentation from the front
panel of the examples' Main VIs into standalone README files.^ This is
mainly for the benefit of visitors who browse the project directly on
GitHub, as the README.md files get formatted and rendered nicely there.

The front panel of the Main VIs are still the more intuitive place for
users to find such documentation, so there is no need to distribute
most^ of these README files in the release package.

^Exception: The "Porting XControls" example is formatted a bit
differently; its Main VI is formatted to mimics its official XControl-
based counterpart.